feat(change paging for filter in handler and controller)
