+++ src/models/artwork.js
@@ -17,7 +17,7 @@ module.exports = (sequelize) => {
         type: DataTypes.STRING,
         allowNull: false,
       },
-      imageURL: {
+      image: {
         type: DataTypes.STRING,
         allowNull: false,
       },
