+++ src/routes/artworksRouter.js
@@ -5,7 +5,7 @@ const upload = require('../utils/cloudinary');
 const getArtworkHandler = require('../handlers/artworks/getHandler');
 const getArtIdHandler = require('../handlers/artworks/getByIDHandler');
 const postArtworkHandler = require('../handlers/artworks/postHandler');
-const artworksPagingHandler = require('../handlers/artworks/pagingHandler');
+const artworksFilterHandler = require('../handlers/artworks/filterHandler');
 const putArtworkHandler = require('../handlers/artworks/putArtworkHandler');
 const deleteArtworkHandler = require('../handlers/artworks/deleteHandler');
 
@@ -13,7 +13,7 @@ artworksRouter.get('/', getArtworkHandler);
 
 artworksRouter.get('/detail/:id', getArtIdHandler);
 
-artworksRouter.get('/db', artworksPagingHandler);
+artworksRouter.get('/db', artworksFilterHandler);
 
 artworksRouter.post('/', upload.single('image'), postArtworkHandler);
