+++ src/routes/usersRouter.js
@@ -6,11 +6,14 @@ const getUserIdHandler = require('../handlers/users/getByIDHandler');
 const loginHandler = require('../handlers/users/loginHandler');
 const updateUsersHandler = require('../handlers/users/updateHandler');
 const authenticateToken = require('../utils/authenticateToken'); // Importa el middleware
+const getFavoritesHandler = require('../handlers/users/getFavoritesHandler');
 
 usersRouter.post('/', postUsersHandler);
 usersRouter.get('/', getUsersHandler);
 usersRouter.get('/detail/:id', getUserIdHandler);
 usersRouter.put('/edit', authenticateToken, updateUsersHandler);
 usersRouter.post('/login', loginHandler);
+
+usersRouter.get('/favorites/:id', getFavoritesHandler);
 
 module.exports = usersRouter;
